http_conn_validator: Initial commit
Initial commit of the new provider http_conn_validator that allows
one to check if a web service is actually listening on a given ip,
port and specific URL. The https protocol can be used if specified.
It supports both IPv4 and IPv6 address.

It either takes a string connection in the namevar or received the
informations through the `server`, `port`, `use_ssl` and `test_url`
parameters.

Case 1:

   http_conn_validator { 'https://127.0.0.1:8080/test_url' : }

Case 2:

   http_conn_validator { 'local elasticsearch' :
     server => '127.0.0.1',
     port   => 9200,
   }